pyproject.toml
[project]
name = "menage2"
version = "0.1.0"
description = ""
authors = [{ name = "Christian Theune", email = "christian@theune.cc" }]
requires-python = "~=3.10.0"
readme = "README.md"
license.text = "MIT"
dependencies = [
    "alembic>=1.11.1",
    "argon2-cffi>=23.1.0",
    "arrow>=1.3.0",
    "babel>=2.12.1",
    "cachecontrol>=0.13.1",
    "filelock>=3.13.1",
    "httplib2>=0.22.0",
    "markdown>=3.7",
    "peppercorn",
    "plaster-pastedeploy>=1.0.1",
    "psycopg2-binary>=2.9.9",
    "webauthn>=2.1.0",
    "pyramid-chameleon>=0.3",
    "pyramid-debugtoolbar>=4.10",
    "pyramid-beaker>=0.8",
    "beaker>=1.12.1",
    "pyramid-retry>=2.1.1",
    "pyramid-tm>=2.5",
    "pyramid>=2.0.1",
    "pytz>=2023.3",
    "rtmapi>=0.7.2",
    "sqlalchemy>=2.0.19",
    "transaction>=3.1.0",
    "waitress>=2.1.2",
    "zope-sqlalchemy~=3.0",
]

[project.scripts]
initialize_menage2_db = "menage2.scripts.initialize_db:main"
set_ingredient_tags = "menage2.scripts.ingredients:main"

[project.entry-points."paste.app_factory"]
main = "menage2:main"

[dependency-groups]
dev = [
    "fakeredis>=2.20.0",
    "webtest>=3.0.0,<4",
    "pytest>=7.4.0,<8",
    "pytest-cov>=4.1.0,<5",
    "playwright>=1.40.0",
    "pytest-playwright>=0.4.0",
    "sqlalchemy-stubs"
]

[tool.uv.sources]
peppercorn = { git = "https://github.com/ctheune/peppercorn.git", branch = "htmx-workaround-1686-wrong-form-dataset-construction" }

[build-system]
requires = ["setuptools>=42"]
build-backend = "setuptools.build_meta"
